Replace `Players` collection
The game uses a collection of slot to store the players of the current
layer. When addressing the players by index, the index point to a slot.
Since the slot are reserved when a player is removed from the collection
the other player aren't shuffled around (their index don't change).
This collection implement that mechanism